Made change per Chasinglogic's recommendation
Note: LOCALAPPDATA will return  "C:\Users\<username>\AppData\Local". I'm opting for Local over Roaming (APPDATA) for privacy purposes. That is, if APPDATA is used and the machine is connected to a domain where Roaming Profiles is configured/enabled, taskforge files would be synced to the organization's fileshare and open to its administration/inspection.